Slate board
• “SLATE BOARD” is a user oriented software that is concerned of the
users entertainment and provide other functions such as, file transfer,
Buffering MP3 Player, Various Filtering Technique, Image Flipping,
Reorder Image Size, Voice Chat, Text Chat, Open and Save image. It is an
integrated technology that can provide multitasking of different
features. Thus it is basically developed keeping in mind the needs of
the user and providing key functions as well as full proof
entertainment. It is a system messenger type communicational and an
entertainment device and provides a high level of user friendliness.